refactor: improve code layout to prepare for chatbot
