+++ src/aind_metadata_viz/query.py
@@ -1,17 +1,14 @@
 """App for generating metadata queries"""
-import os
-from typing import Optional
-import json
-
 import pandas as pd
 import panel as pn
 import param
 
-from aind_metadata_viz.utils import outer_style, AIND_COLORS, sort_with_none
+from aind_metadata_viz.utils import outer_style, AIND_COLORS, FIXED_WIDTH
+from aind_metadata_viz.query.simple_query import QueryBuilder
+from aind_metadata_viz.query.viewer import QueryViewer
+from aind_metadata_viz.query.database import get_docdb_records
 
-from aind_data_access_api.document_db import MetadataDbClient
 
-FIXED_WIDTH = 1200
 background_param = pn.state.location.query_params.get("background", "dark_blue")
 background_color = AIND_COLORS.get(background_param, AIND_COLORS["dark_blue"])
 
@@ -23,361 +20,6 @@ body {{
 }}
 """
 pn.config.raw_css.append(css)
-
-
-API_GATEWAY_HOST = os.getenv("API_GATEWAY_HOST", "api.allenneuraldynamics-test.org")
-DATABASE = os.getenv("DATABASE", "metadata_index")
-COLLECTION = os.getenv("COLLECTION", "data_assets")
-
-docdb_api_client = MetadataDbClient(
-    host=API_GATEWAY_HOST,
-    database=DATABASE,
-    collection=COLLECTION,
-)
-
-DF_KEYS = ["name"]
-
-
-# Helpers to get option lists
-@pn.cache(ttl=86400)  # Cache for 24 hours
-def get_project_names():
-    project_options = docdb_api_client.aggregate_docdb_records(
-        pipeline=[
-            {
-                "$group": {
-                    "_id": "$data_description.project_name"
-                }
-            },
-            {
-                "$sort": {
-                    "_id": 1  # Optional: sorts alphabetically
-                }
-            }
-        ]
-    )
-    project_options = [project["_id"] for project in project_options]
-
-    if project_options:
-        project_options = sort_with_none(project_options)
-    return project_options
-
-
-@pn.cache(ttl=86400)  # Cache for 24 hours
-def get_subject_ids(project_name: Optional[str]):
-    """Get subject IDs"""
-
-    if not project_name:
-        return []
-
-    subject_options = docdb_api_client.aggregate_docdb_records(
-        pipeline=[
-            {  # filter by project name
-                "$match": {
-                    "data_description.project_name": project_name
-                }
-            },
-            {
-                "$group": {
-                    "_id": "$subject.subject_id"
-                }
-            },
-        ]
-    )
-    subject_options = [subject["_id"] for subject in subject_options]
-    if subject_options:
-        subject_options = sort_with_none(subject_options)
-    return subject_options
-
-
-@pn.cache(ttl=86400)  # Cache for 24 hours
-def get_modalities(project_name: Optional[str]):
-    """Get modality abbreviations"""
-
-    if not project_name:
-        return []
-
-    modality_options = docdb_api_client.aggregate_docdb_records(
-        pipeline=[
-            {  # filter by project name
-                "$match": {
-                    "data_description.project_name": project_name
-                }
-            },
-            {
-                "$unwind": "$data_description.modality"
-            },
-            {
-                "$group": {
-                    "_id": "$data_description.modality.abbreviation"
-                }
-            },
-        ]
-    )
-    modality_options = [modality["_id"] for modality in modality_options]
-    if modality_options:
-        modality_options = sort_with_none(modality_options)
-    return modality_options
-
-
-@pn.cache(ttl=86400)  # Cache for 24 hours
-def get_session_types(project_name: Optional[str]):
-    """Get session types"""
-
-    if not project_name:
-        return []
-
-    session_type_options = docdb_api_client.aggregate_docdb_records(
-        pipeline=[
-            {  # filter by project name
-                "$match": {
-                    "data_description.project_name": project_name
-                }
-            },
-            {
-                "$group": {
-                    "_id": "$session.session_type"
-                }
-            },
-        ]
-    )
-    session_type_options = [session["_id"] for session in session_type_options]
-    if session_type_options:
-        session_type_options = sort_with_none(session_type_options)
-    return session_type_options
-
-
-@pn.cache(ttl=60*60)  # Cache for 1 hour
-def get_docdb_records(filter_query: dict):
-    """Get a set of records"""
-    return docdb_api_client.retrieve_docdb_records(
-        filter_query=filter_query,
-        projection={key: 1 for key in DF_KEYS},
-    )
-
-
-class QueryViewer(param.Parameterized):
-    """Class for displaying the result of a query"""
-
-    query = param.Dict(default={}, allow_None=True)
-
-    def __init__(self, query, **params):
-        super().__init__(**params)
-        self.query = query
-        self.query_pane = pn.pane.JSON(
-            object=self.query,
-            width=FIXED_WIDTH-50,
-        )
-
-        self.hidden_html = pn.pane.HTML(
-            object="",
-            width=0,
-            height=0,
-        )
-
-        self.copy_button = pn.widgets.Button(
-            name="",
-            icon="copy",
-            button_type="primary",
-            width=40,
-            height=30,
-        )
-        self.copy_button.on_click(self.copy_to_clipboard)
-
-    def update(self, query: dict):
-        """Update the query pane with a new query"""
-        self.query = query
-        self.query_pane.object = query
-
-    def copy_to_clipboard(self, event):
-        """Copy the query to clipboard"""
-        query_data = self.query.copy()
-        if "_name" in query_data:
-            del query_data["_name"]
-
-        clipboard_js = f"""
-        <script>
-        navigator.clipboard.writeText('{json.dumps(query_data)}');
-        </script>
-        """
-        self.hidden_html.object = clipboard_js
-        self.hidden_html.object = ""
-
-    def panel(self):
-        """Return the query viewer panel"""
-        return pn.Row(
-            self.hidden_html,
-            pn.Column(self.query_pane, width=FIXED_WIDTH-150),
-            pn.Column(self.copy_button, align='end'),
-            width=FIXED_WIDTH-50
-        )
-
-
-class QueryBuilder(param.Parameterized):
-    """Class for generating simple metadata queries"""
-
-    project_name = param.String(default="", allow_None=True)
-    subject_ids = param.List(default=[], allow_None=True)
-    modalities = param.List(default=[], allow_None=True)
-    session_types = param.List(default=[], allow_None=True)
-    queries = param.List(default=[])
-
-    def __init__(self, **params):
-        super().__init__(**params)
-        self.query_viewer = QueryViewer({})
-        self.project_name_selector = pn.widgets.Select(
-            name="DataDescription.project_name",
-            options=[""] + get_project_names(),  # Add empty string option
-            value=self.project_name,
-        )
-        self.subject_id_selector = pn.widgets.MultiChoice(
-            name="Subject.subject_id",
-            options=[""] + get_subject_ids(None),  # Add empty string option
-            value=self.subject_ids,
-            disabled=True,
-        )
-        self.modality_selector = pn.widgets.MultiChoice(
-            name="DataDescription.modality",
-            options=[""] + get_modalities(None),  # Add empty string option
-            value=self.modalities,
-            disabled=True,
-        )
-        self.session_type_selector = pn.widgets.MultiChoice(
-            name="Session.session_type",
-            options=[""] + get_session_types(None),  # Add empty string option
-            value=[],
-            width=500,
-            disabled=True,
-        )
-
-        self.project_name_selector.link(self, value="project_name")
-        self.subject_id_selector.link(self, value="subject_ids")
-        self.modality_selector.link(self, value="modalities")
-        self.session_type_selector.link(self, value="session_types")
-
-        self.query_button = pn.widgets.Button(
-            name="Submit query",
-            button_type="primary",
-        )
-        pn.bind(self.save_query, self.query_button, watch=True)
-
-    def options_panel(self):
-        """Create the options panel for the query"""
-
-        selector_col = pn.Column(
-            pn.Row(
-                self.project_name_selector,
-                self.subject_id_selector,
-                self.modality_selector,
-            ),
-            pn.Row(
-                self.session_type_selector,
-            ),
-            width=FIXED_WIDTH-150,
-        )
-
-        submit_col = pn.Column(
-            self.query_button,
-            width=100,
-        )
-
-        return pn.Row(selector_col, submit_col)
-
-    @pn.depends("project_name", watch=True)
-    def update_subject_id_options(self):
-        """Clear the subject ID value and change options"""
-
-        if self.project_name != "":
-            # update subject ID options
-            self.subject_id_selector.options = get_subject_ids(self.project_name)
-            self.subject_id_selector.disabled = False
-        else:
-            # reset subject ID options
-            self.subject_id_selector.options = []
-            self.subject_id_selector.disabled = True
-
-        self.subject_id_selector.value = []
-        self.subject_ids = []
-
-    @pn.depends("project_name", watch=True)
-    def update_modality_options(self):
-        """Clear the modality value and change options"""
-
-        if self.project_name != "":
-            # update modality options
-            self.modality_selector.options = get_modalities(self.project_name)
-            self.modality_selector.disabled = False
-        else:
-            # reset modality options
-            self.modality_selector.options = []
-            self.modality_selector.disabled = True
-
-        self.modality_selector.value = []
-        self.modalities = []
-
-    @pn.depends("project_name", watch=True)
-    def update_session_type_options(self):
-        """Clear the session type value and change options"""
-
-        if self.project_name != "":
-            # update session type options
-            self.session_type_selector.options = get_session_types(self.project_name)
-            self.session_type_selector.disabled = False
-        else:
-            # reset session type options
-            self.session_type_selector.options = []
-            self.session_type_selector.disabled = True
-
-        self.session_type_selector.value = []
-        self.session_types = []
-
-    @pn.depends("project_name", "subject_ids", "modalities", "session_types", watch=True)
-    def update_query_panel(self):
-        """Update the query panel content dynamically"""
-        self.query_button.disabled = False
-        query_dict = {"_name": f"Query {len(self.queries) + 1}"}
-
-        if self.project_name != "":
-            query_dict["data_description.project_name"] = self.project_name
-
-        if self.subject_ids != []:
-            query_dict["subject.subject_id"] = {
-                "$in": self.subject_ids
-            }
-
-        if self.modalities != []:
-            query_dict["data_description.modality.abbreviation"] = {
-                "$in": self.modalities
-            }
-
-        if self.session_types != []:
-            query_dict["session.session_type"] = {
-                "$in": self.session_types
-            }
-
-        self.query_viewer.update(query_dict)
-
-        if len(query_dict.keys()) <= 1:
-            self.query_button.disabled = True
-            self.query_button.name = "Cannot submit empty query"
-            self.query_button.button_type = "danger"
-        else:
-            self.query_button.name = "Submit query"
-            self.query_button.disabled = False
-            self.query_button.button_type = "primary"
-
-    def save_query(self, event):
-        """Store the current query in the queries list"""
-        self.queries = self.queries + [self.query_viewer.query_pane.object]
-        self.query_button.disabled = True
-
-    def panel(self):
-        """Return the full panel"""
-        return pn.Column(
-            self.options_panel(),
-            pn.pane.Markdown("## Query"),
-            self.query_viewer.panel(),
-            width=FIXED_WIDTH,
-        )
 
 
 class QueryResult(param.Parameterized):
