+++ src/aind_metadata_viz/query/chat_query.py
@@ -1,0 +1,1 @@
+"""Complex query generation via chatbot"""
